refactor(simplicity): flatten speculative abstractions and reduce cognitive complexity (closes #135)
