+++ lib/materials/index.ts
@@ -1,11 +1,9 @@
 export * from './chunker';
 export * from './concept-extraction';
-export * from './deletion';
-
 export * from './ingestion';
 export * from './inspection';
 export * from './intake';
-export * from './purge';
+export * from './lifecycle';
 export * from './rasterizer';
 export * from './retrieval';
 export * from './types';
