+++ lib/utils.ts
@@ -9,11 +9,7 @@ export function cn(...inputs: ClassValue[]): string {
 }
 
 export function generateUUID(): string {
-  return 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, (c) => {
-    const r = (Math.random() * 16) | 0;
-    const v = c === 'x' ? r : (r & 0x3) | 0x8;
-    return v.toString(16);
-  });
+  return crypto.randomUUID();
 }
 
 export function convertToUIMessages(messages: DBMessage[]): ChatMessage[] {
